refactor getExtensionArn and getLayerArn to support ARM, move functions to commons.ts and also its tests to commons.test.ts
also upgraded the aws-sdk to properly get the Architectures field in FunctionConfiguration from Lambda
