+++ package.json
@@ -41,7 +41,7 @@
   },
   "dependencies": {
     "async-retry": "1.3.1",
-    "aws-sdk": "2.903.0",
+    "aws-sdk": "2.1012.0",
     "axios": "0.21.2",
     "chalk": "3.0.0",
     "clipanion": "2.2.2",
