+++ src/commands/lambda/functions/instrument.ts
@@ -5,11 +5,9 @@ import {
   CI_KMS_API_KEY_ENV_VAR,
   CI_SITE_ENV_VAR,
   DD_LAMBDA_EXTENSION_LAYER_NAME,
-  DEFAULT_LAYER_AWS_ACCOUNT,
   ENVIRONMENT_ENV_VAR,
   EXTRA_TAGS_ENV_VAR,
   FLUSH_TO_LOG_ENV_VAR,
-  GOVCLOUD_LAYER_AWS_ACCOUNT,
   HANDLER_LOCATION,
   KMS_API_KEY_ENV_VAR,
   LAMBDA_HANDLER_ENV_VAR,
@@ -17,7 +15,6 @@ import {
   LOG_LEVEL_ENV_VAR,
   MERGE_XRAY_TRACES_ENV_VAR,
   Runtime,
-  RUNTIME_LAYER_LOOKUP,
   SERVICE_ENV_VAR,
   SITE_ENV_VAR,
   TRACE_ENABLED_ENV_VAR,
@@ -26,7 +23,14 @@ import {
 import {FunctionConfiguration, InstrumentationSettings, LogGroupConfiguration, TagConfiguration} from '../interfaces'
 import {calculateLogGroupUpdateRequest} from '../loggroup'
 import {calculateTagUpdateRequest} from '../tags'
-import {addLayerARN, getLambdaFunctionConfigs, isLambdaActive, isSupportedRuntime} from './commons'
+import {
+  addLayerARN,
+  getExtensionArn,
+  getLambdaFunctionConfigs,
+  getLayerArn,
+  isLambdaActive,
+  isSupportedRuntime,
+} from './commons'
 
 export const getFunctionConfigs = async (
   lambda: Lambda,
@@ -61,8 +65,8 @@ export const getFunctionConfig = async (
   }
 
   await isLambdaActive(lambda, config, functionARN)
-  const lambdaLibraryLayerArn: string = getLayerArn(runtime, settings, region)
-  const lambdaExtensionLayerArn: string = getExtensionArn(settings, region)
+  const lambdaLibraryLayerArn: string = getLayerArn(config, region, settings)
+  const lambdaExtensionLayerArn: string = getExtensionArn(config, region, settings)
   const updateRequest = calculateUpdateRequest(
     config,
     settings,
@@ -126,28 +130,6 @@ export const getLambdaConfigsFromRegEx = async (
   }
 
   return functionsToUpdate
-}
-
-export const getLayerArn = (runtime: Runtime, settings: InstrumentationSettings, region: string) => {
-  const layerName = RUNTIME_LAYER_LOOKUP[runtime]
-  const account = settings.layerAWSAccount ?? DEFAULT_LAYER_AWS_ACCOUNT
-  const isGovCloud = region.startsWith('us-gov')
-  if (isGovCloud) {
-    return `arn:aws-us-gov:lambda:${region}:${GOVCLOUD_LAYER_AWS_ACCOUNT}:layer:${layerName}`
-  }
-
-  return `arn:aws:lambda:${region}:${account}:layer:${layerName}`
-}
-
-export const getExtensionArn = (settings: InstrumentationSettings, region: string) => {
-  const layerName = DD_LAMBDA_EXTENSION_LAYER_NAME
-  const account = settings.layerAWSAccount ?? DEFAULT_LAYER_AWS_ACCOUNT
-  const isGovCloud = region.startsWith('us-gov')
-  if (isGovCloud) {
-    return `arn:aws-us-gov:lambda:${region}:${GOVCLOUD_LAYER_AWS_ACCOUNT}:layer:${layerName}`
-  }
-
-  return `arn:aws:lambda:${region}:${account}:layer:${layerName}`
 }
 
 export const calculateUpdateRequest = (
@@ -272,7 +254,7 @@ export const calculateUpdateRequest = (
       newEnvVars[KMS_API_KEY_ENV_VAR] === undefined
     ) {
       throw new Error(
-        `When 'extensionLayer' is set, ${CI_API_KEY_ENV_VAR} or ${CI_KMS_API_KEY_ENV_VAR} must also be set`
+        `When 'extensionLayer' is set, ${CI_API_KEY_ENV_VAR} or ${CI_KMS_API_KEY_ENV_VAR} must also be set.`
       )
     }
   })
